// src/core/field-engine/core/input-base/field-input-base-types.ts
export interface IField extends IFieldDescriptor {
    [key: string]: any
    id: number
    name: string
    label: string

    enabled: boolean
    isValid: boolean
    isDirty: boolean
    isPristine: boolean
    isFocus: boolean

    value: FieldDataTypes
    originalValue: FieldDataTypes

    /** message helper method : uses treacker and fallbacks to console */
    message: (type: TrackingType, source: string, message: string) => void
    /** Core methods */
    setValue: (value: FieldDataTypes) => void
    getValue: () => FieldDataTypes

    setFocus: () => void
    clear: () => void
    enable: (enabled: boolean) => void

    hasChanges: (callback: () => void) => void
    handleValidation: <T extends IEvents>(event?: T, data?: IValidationStrategyData) => void

    handleOnBlur: <T extends IEvents>(data?: T) => void
    handleOnFocus: <T extends IEvents>(data?: T) => void
    handleOnClear: <T extends IEvents>(data?: T) => void

    /** Dependency accessors */
    domManager: IDomManager<HTMLInputElement>
    drawer: IDrawerBaseInput
    styleManager: IStyleManager
    notificationManager: INotificationManager
    trackingManager: ITrackingManager
    validationManager: IValidationManager
    vlaueManager: IValueManager
}

export type SchemeToDescriptorConverterType = (scheme: IEntityScheme) => IFieldDescriptor

/** @warning: should not being used outside it's main implementation prefer the use of @Ifield */
export type IFieldInput = IFieldBaseInput & Omit<IFieldDescriptor, 'validationOptions' | 'options'>

export interface IFieldBaseInput extends IField, IInitializableDependency {
    new (descriptor: IFieldDescriptor): IFieldBaseInput

    /** initializer builders */
    initializeProperties: (descriptor: IFieldDescriptor) => void
    checkInitialized: () => IInitilizationCheckResult
    useDomManager: (dommableInstance?: IDomManager<HTMLInputElement>) => IFieldBaseInput
    useNotificationManager: (notifierInstance?: INotificationManager) => IFieldBaseInput
    useTrackingManager: (trackerInstance?: ITrackingManager) => IFieldBaseInput
    useValidationManager: (validationStrategyInstance?: IValidationManager) => IFieldBaseInput
    useValueManager: (valueStrategyInstance?: IValueManager) => IFieldBaseInput
    useDrawerManager: (drawerableInstance?: IDrawerBaseInput) => IFieldBaseInput
    useStyleManager: (stylerInstance?: IStyleManager) => IFieldBaseInput
}